+++ OCRCoreService/Services/OCREngine.cs
@@ -71,6 +71,7 @@ namespace OCRCoreService.Services
             oCRParameter.det_db_box_thresh = 0.618f;
             oCRParameter.visualize = false;
             oCRParameter.return_word_box = _ocrConfig.return_word_box;
+            oCRParameter.ocr_instance_count = _ocrConfig.ocr_instance_count;
 
             para.ocrpara = oCRParameter;
             para.paraType = EnumParaType.Class;